backend/requirements.txt
# RAG Pipeline Optimizer - Backend Dependencies
# Install: pip install -r requirements.txt

# PDF + Image parsing
pymupdf==1.24.11

# Web framework
fastapi==0.128.0
uvicorn==0.40.0
python-multipart==0.0.21
pydantic==2.12.5

# LangChain ecosystem
langchain==1.2.6
langchain-community==0.4.1
langchain-huggingface==1.2.0
langchain-ollama==1.0.1

# Vector store
chromadb==1.4.1

# Embeddings & LLM
sentence-transformers==5.2.0
numpy

# Supabase
supabase==2.27.2

# Auth
PyJWT==2.10.1

# HTTP & networking
requests==2.32.5
httpx==0.28.1

# Config
python-dotenv==1.2.1
